Reject package with embeddedReadme file with featureFlag
