+++ src/NuGetGallery.Core/Packaging/PackageMetadataStrings.cs
@@ -21,5 +21,6 @@ namespace NuGetGallery.Packaging
         public const string Tags = "tags";
         public const string Title = "title";
         public const string Icon = "icon";
+        public const string Readme = "readme";
     }
 }
