+++ src/NuGetGallery/App_Start/DefaultDependenciesModule.cs
@@ -345,6 +345,11 @@ namespace NuGetGallery
                 .As<ISymbolPackageService>()
                 .InstancePerLifetimeScope();
 
+            builder.RegisterType<PackageMetadataValidationService>()
+                .AsSelf()
+                .As<IPackageMetadataValidationService>()
+                .InstancePerLifetimeScope();
+
             builder.RegisterType<PackageUploadService>()
                 .AsSelf()
                 .As<IPackageUploadService>()
