+++ tests/NuGetGallery.Core.Facts/Packaging/PackageMetadataFacts.cs
@@ -3,11 +3,9 @@
 
 using System;
 using System.Collections.Generic;
-using System.Globalization;
 using System.IO;
 using System.IO.Compression;
 using System.Linq;
-using System.Threading;
 using NuGet.Packaging;
 using NuGet.Packaging.Core;
 using NuGet.Versioning;
