+++ tests/NuGetGallery.Core.Facts/TestUtils/TestPackage.cs
@@ -43,7 +43,8 @@ namespace NuGetGallery
             Func<string> getCustomNodes = null,
             string licenseExpression = null,
             string licenseFilename = null,
-            string iconFilename = null)
+            string iconFilename = null,
+            string readmeFilename =null)
         {
             var fullNuspec = (@"<?xml version=""1.0""?>
                 <package xmlns=""http://schemas.microsoft.com/packaging/2011/08/nuspec.xsd"">
@@ -64,6 +65,7 @@ namespace NuGetGallery
                         <licenseUrl>" + (licenseUrl?.AbsoluteUri ?? string.Empty) + @"</licenseUrl>
                         " + WriteLicense(licenseExpression, licenseFilename) + @"
                         " + WriteIcon(iconFilename) + @"
+                        " + WriteReadme(readmeFilename) + @"
                         <projectUrl>" + (projectUrl?.ToString() ?? string.Empty) + @"</projectUrl>
                         <iconUrl>" + (iconUrl?.ToString() ?? string.Empty) + @"</iconUrl>
                         <packageTypes>" + WritePackageTypes(packageTypes) + @"</packageTypes>
@@ -111,6 +113,16 @@ namespace NuGetGallery
             if (iconFilename != null)
             {
                 return $"<icon>{iconFilename}</icon>";
+            }
+
+            return string.Empty;
+        }
+
+        private static string WriteReadme(string readmeFilename)
+        {
+            if (readmeFilename != null)
+            {
+                return $"<readme>{readmeFilename}</readme>";
             }
 
             return string.Empty;
@@ -215,7 +227,9 @@ namespace NuGetGallery
             string licenseFilename = null,
             byte[] licenseFileContents = null,
             string iconFilename = null,
-            byte[] iconFileContents = null)
+            byte[] iconFileContents = null,
+            string readmeFilename = null,
+            byte[] readmeFileContents = null)
         {
             return CreateTestPackageStream(packageArchive =>
             {
@@ -225,11 +239,12 @@ namespace NuGetGallery
                     WriteNuspec(stream, true, id, version, title, summary, authors, owners, description, tags, language,
                         copyright, releaseNotes, minClientVersion, licenseUrl, projectUrl, iconUrl,
                         requireLicenseAcceptance, developmentDependency, packageDependencyGroups, packageTypes, isSymbolPackage, repositoryMetadata,
-                        getCustomNuspecNodes, licenseExpression, licenseFilename, iconFilename);
+                        getCustomNuspecNodes, licenseExpression, licenseFilename, iconFilename, readmeFilename);
                 }
 
                 licenseFilename = AddBinaryFile(packageArchive, licenseFilename, licenseFileContents);
                 iconFilename = AddBinaryFile(packageArchive, iconFilename, iconFileContents);
+                readmeFilename = AddBinaryFile(packageArchive, readmeFilename, readmeFileContents);
 
                 if (populatePackage != null)
                 {
